Fixed bug in image.cpp that could cause a scaled dimension to become zero, causing an error in the scaler.
Made cancelled atomic, and made sure that we atomically get a copy of
the request shared_ptr in the methods that are called from the thread pools,
so we don't read stale memory. Made a bunch of immutable handler members const,
for clarity.

More aggressive testing.

Fixed album_art test to force a cache miss on each iteration.